Chapter 1 homework solved
There was a n erro while testing the interface.js using test.js. The
issue was with the fact that I was calling the callback after checking
for error. Apperently, it was not suppose to be that way. I was suppose
to call the callback directly and all the nullability of error is
checked at the place where callback is used.

Also there are files for chapter 2 added.